# app/app/api/auth/routes.py
# app/api/auth/routes.py
"""
Controller unificado de autenticação.
Combina funcionalidades de auth_controller.py e auth_ui_controller.py
"""
from __future__ import annotations

from datetime import datetime
from typing import Optional

from fastapi import APIRouter, Request, Response, Body, Cookie, Header
from fastapi.responses import JSONResponse

from app.core.logging_config import get_logger
from app.services.auth_wiring import build_auth_service_from_settings
from app.services.auth_service import AuthError
from app.config import settings

from .schemas import (
    LoginRequest, LoginResponse, RefreshResponse, LogoutResponse,
    RequestPasswordResetRequest, RequestPasswordResetResponse,
    ConfirmPasswordResetRequest, ConfirmPasswordResetResponse,
    ErrorResponse, ApiKeyValidationResponse, User, build_user_response
)
from .dependencies import verify_api_key, require_user, CurrentUser

logger = get_logger("auth")

# Router principal de autenticação
router = APIRouter(prefix="/v1/auth", tags=["Authentication"])

# Serviço de autenticação
auth_service = build_auth_service_from_settings()


# =========================
#   HELPERS
# =========================
def _error_response(status_code: int, *, error: str, message: str, details: Optional[dict] = None,
                    trace_id: Optional[str] = None) -> JSONResponse:
    payload = ErrorResponse(error=error, message=message, details=details, traceId=trace_id).model_dump(exclude_none=True)
    return JSONResponse(status_code=status_code, content=payload)


def _client_hints(req: Request) -> tuple[Optional[str], Optional[str]]:
    """Captura User-Agent e IP do cliente."""
    ua = req.headers.get("User-Agent")
    fwd = req.headers.get("X-Forwarded-For")
    ip = (fwd.split(",")[0].strip() if fwd else req.client.host if req.client else None)
    return ua, ip





# =========================
#   ENDPOINTS PÚBLICOS (UI)
# =========================
@router.post(
    "/login",
    response_model=LoginResponse,
    responses={
        400: {"model": ErrorResponse},
        401: {"model": ErrorResponse},
        500: {"model": ErrorResponse},
    },
)
async def login(
    req: Request,
    response: Response,
    body: LoginRequest = Body(...),
) -> LoginResponse | JSONResponse:
    """Login de usuário - retorna JWT tokens."""
    logger.info("auth.login.start", extra={"login": body.login})
    try:
        ua, ip = _client_hints(req)

        bundle = await auth_service.login(
            login=body.login,
            senha=body.senha,
            user_agent=ua,
            client_ip=ip,
        )

        user = build_user_response(bundle.user)

        # Seta cookie HttpOnly + Secure com o refresh
        if bundle.refresh_token:
            response.set_cookie(
                key="refreshToken",
                value=bundle.refresh_token,
                httponly=True,
                secure=True,
                samesite="strict",
                path="/v1/auth",
                max_age=bundle.refresh_expires_in,
                domain=settings.COOKIE_DOMAIN,
            )

        logger.info("auth.login.ok", extra={"login": body.login})
        return LoginResponse(
            accessToken=bundle.access_token,
            accessTokenExpiresIn=bundle.access_expires_in,
            refreshToken=bundle.refresh_token,
            refreshTokenExpiresIn=bundle.refresh_expires_in,
            user=user,
        )

    except AuthError as e:
        logger.info("auth.login.fail", extra={"login": body.login, "code": e.code})
        status_map = {
            "ERR_USER_NOT_FOUND": 401,
            "ERR_INVALID_SENHA": 401,
            "ERR_USER_DISABLED": 403,
        }
        sc = status_map.get(e.code, 400)
        return _error_response(sc, error=e.code, message=e.message)
    except Exception:
        logger.exception("auth.login.error")
        return _error_response(500, error="ERR_INTERNAL", message="Erro interno")


@router.post(
    "/refresh",
    response_model=RefreshResponse,
    responses={
        400: {"model": ErrorResponse},
        401: {"model": ErrorResponse},
        403: {"model": ErrorResponse},
        500: {"model": ErrorResponse},
    },
)
async def refresh(
    req: Request, 
    response: Response, 
    refresh_token: Optional[str] = Cookie(None, alias="refreshToken")
) -> RefreshResponse | JSONResponse:
    """Refresh de token JWT."""
    logger.info("auth.refresh.start")
    try:
        ua, ip = _client_hints(req)

        # Fallback: se não veio cookie, tenta o header legado
        raw_refresh = refresh_token or req.headers.get("X-Refresh-Token")

        bundle = await auth_service.refresh_token(
            raw_refresh_token=raw_refresh, user_agent=ua, client_ip=ip
        )
        logger.info("auth.refresh.ok")

        # Renova o cookie com o NOVO refresh
        if bundle.refresh_token:
            response.set_cookie(
                key="refreshToken",
                value=bundle.refresh_token,
                httponly=True,
                secure=True,
                samesite="strict",
                path="/v1/auth",
                max_age=bundle.refresh_expires_in,
                domain=settings.COOKIE_DOMAIN,
            )

        return RefreshResponse(
            accessToken=bundle.access_token,
            accessTokenExpiresIn=bundle.access_expires_in,
            refreshToken=bundle.refresh_token,
            refreshTokenExpiresIn=bundle.refresh_expires_in,
        )

    except AuthError as e:
        logger.info("auth.refresh.fail", extra={"code": e.code})
        status_map = {
            "ERR_REFRESH_MISSING": 401,
            "ERR_REFRESH_INVALID": 401,
            "ERR_REFRESH_REVOKED": 401,
            "ERR_REFRESH_EXPIRED": 401,
            "ERR_USER_DISABLED": 403,
            "ERR_USER_NOT_FOUND": 401,
        }
        sc = status_map.get(e.code, 400)
        return _error_response(sc, error=e.code, message=e.message)
    except Exception:
        logger.exception("auth.refresh.error")
        return _error_response(500, error="ERR_INTERNAL", message="Erro interno")


@router.post(
    "/logout",
    response_model=LogoutResponse,
    responses={
        400: {"model": ErrorResponse},
        401: {"model": ErrorResponse},
        500: {"model": ErrorResponse},
    },
)
async def logout(
    req: Request,
    response: Response,
    refresh_token: Optional[str] = Cookie(None, alias="refreshToken"),
) -> LogoutResponse | JSONResponse:
    """Logout de usuário - invalida refresh token."""
    logger.info("auth.logout.start")
    try:
        # Fallback: se não veio cookie, tenta header legado
        raw_refresh = refresh_token or req.headers.get("X-Refresh-Token")

        await auth_service.logout(raw_refresh_token=raw_refresh)
        logger.info("auth.logout.ok")

        # Apaga o cookie
        response.delete_cookie(
            key="refreshToken",
            path="/v1/auth",
            domain=settings.COOKIE_DOMAIN
        )

        return LogoutResponse(ok=True)

    except Exception:
        logger.exception("auth.logout.error")
        return _error_response(500, error="ERR_INTERNAL", message="Erro interno")


# =========================
#   RESET DE SENHA
# =========================
@router.post(
    "/request-password-reset",
    response_model=RequestPasswordResetResponse,
    responses={
        400: {"model": ErrorResponse},
        401: {"model": ErrorResponse},
        500: {"model": ErrorResponse},
    },
)
async def request_password_reset(
    req: Request,
    body: RequestPasswordResetRequest,
) -> RequestPasswordResetResponse | JSONResponse:
    """Solicita reset de senha - gera token temporário."""
    logger.info("auth.request_password_reset.start", extra={"login": body.login})
    try:
        ua, ip = _client_hints(req)

        reset_token = await auth_service.request_password_reset(
            login=body.login,
            user_agent=ua,
            client_ip=ip,
        )

        # TODO: Enviar token por email/SMS
        # Por enquanto, apenas logamos o token (em produção, NUNCA fazer isso!)
        logger.info("auth.request_password_reset.token_generated", extra={
            "login": body.login, 
            "token": reset_token  # REMOVER EM PRODUÇÃO!
        })

        logger.info("auth.request_password_reset.ok", extra={"login": body.login})
        return RequestPasswordResetResponse()

    except AuthError as e:
        logger.info("auth.request_password_reset.fail", extra={"login": body.login, "code": e.code})
        status_map = {
            "ERR_USER_NOT_FOUND": 401,
            "ERR_USER_DISABLED": 403,
        }
        sc = status_map.get(e.code, 400)
        return _error_response(sc, error=e.code, message=e.message)
    except Exception:
        logger.exception("auth.request_password_reset.error")
        return _error_response(500, error="ERR_INTERNAL", message="Erro interno")


@router.post(
    "/confirm-password-reset",
    response_model=ConfirmPasswordResetResponse,
    responses={
        400: {"model": ErrorResponse},
        401: {"model": ErrorResponse},
        500: {"model": ErrorResponse},
    },
)
async def confirm_password_reset(
    req: Request,
    body: ConfirmPasswordResetRequest,
) -> ConfirmPasswordResetResponse | JSONResponse:
    """Confirma reset de senha com token válido."""
    logger.info("auth.confirm_password_reset.start")
    try:
        ua, ip = _client_hints(req)

        await auth_service.confirm_password_reset(
            reset_token=body.token,
            new_password=body.newPassword,
            user_agent=ua,
            client_ip=ip,
        )

        logger.info("auth.confirm_password_reset.ok")
        return ConfirmPasswordResetResponse()

    except AuthError as e:
        logger.info("auth.confirm_password_reset.fail", extra={"code": e.code})
        status_map = {
            "ERR_RESET_TOKEN_MISSING": 400,
            "ERR_RESET_TOKEN_INVALID": 401,
            "ERR_RESET_TOKEN_USED": 401,
            "ERR_RESET_TOKEN_EXPIRED": 401,
            "ERR_USER_NOT_FOUND": 401,
            "ERR_USER_DISABLED": 403,
        }
        sc = status_map.get(e.code, 400)
        return _error_response(sc, error=e.code, message=e.message)
    except Exception:
        logger.exception("auth.confirm_password_reset.error")
        return _error_response(500, error="ERR_INTERNAL", message="Erro interno")


# =========================
#   ENDPOINTS INTERNOS (API KEY)
# =========================
@router.get(
    "/validate",
    response_model=ApiKeyValidationResponse,
    responses={
        401: {"model": ErrorResponse},
    },
)
def validate_api_key(
    x_api_key: str = Header(..., convert_underscores=False)
) -> ApiKeyValidationResponse:
    """Valida API Key para comunicação entre microserviços."""
    verify_api_key(x_api_key)
    return ApiKeyValidationResponse(valid=True)


# =========================
#   JWKS (JSON Web Key Set)
# =========================
@router.get("/.well-known/jwks.json")
def jwks():
    """Endpoint JWKS para validação de JWT."""
    from app.controllers.jwk_controller import jwks as jwks_handler
    return jwks_handler()